pyproject.toml
[tool.poetry]
name = "discollama"
version = "0.1.0"
description = ""
authors = ["Michael Yang <mxyng@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.1"
redis = "^5.0.1"
ollama = "^0.3.2"
discord-py-slash-command = "^4.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 999
indent-width = 2

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]
